Restructured packages of wizard
