+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.measuringview/src/org/palladiosimulator/simulizar/ui/measurementsdashboard/parts/MeasurementsDashboardView.java
@@ -39,7 +39,7 @@ import org.palladiosimulator.measurementsui.abstractviewer.MeasurementsTreeViewe
 import org.palladiosimulator.measurementsui.datamanipulation.ResourceEditor;
 import org.palladiosimulator.measurementsui.datamanipulation.ResourceEditorImpl;
 import org.palladiosimulator.measurementsui.dataprovider.DataApplication;
-import org.palladiosimulator.measurementsui.wizardmain.MeasurementsWizard;
+import org.palladiosimulator.measurementsui.wizard.main.MeasurementsWizard;
 import org.palladiosimulator.measurementsui.wizardmodel.WizardModelType;
 import org.palladiosimulator.monitorrepository.MeasurementSpecification;
 import org.palladiosimulator.monitorrepository.Monitor;
