+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.monitorrepositoryview/META-INF/MANIFEST.MF
@@ -18,7 +18,7 @@ Require-Bundle: org.eclipse.core.runtime,
 Import-Package: org.apache.commons.logging,
  org.apache.log4j,
  org.palladiosimulator.measurementsui.datamanipulation,
- org.palladiosimulator.measurementsui.wizardmain,
+ org.palladiosimulator.measurementsui.wizard.main,
  org.palladiosimulator.measurementsui.wizardmodel
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: monitorrepositoryview,
