+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.monitorrepositoryview/src/org/palladiosimulator/measurementsui/monitorrepositoryview/MonitorRepositorySaveableTreeFormView.java
@@ -3,7 +3,7 @@ package org.palladiosimulator.measurementsui.monitorrepositoryview;
 import org.eclipse.emf.parsley.views.SaveableTreeView;
 import org.eclipse.jface.wizard.WizardDialog;
 import org.eclipse.swt.widgets.Shell;
-import org.palladiosimulator.measurementsui.wizardmain.MeasurementsWizard;
+import org.palladiosimulator.measurementsui.wizard.main.MeasurementsWizard;
 import org.palladiosimulator.measurementsui.wizardmodel.WizardModelType;
 import org.palladiosimulator.monitorrepository.Monitor;
