+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/META-INF/MANIFEST.MF
@@ -3,7 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Wizard
 Bundle-SymbolicName: org.palladiosimulator.measurementsui.wizard;singleton:=true
 Bundle-Version: 1.0.0.qualifier
-Bundle-Activator: org.palladiosimulator.measurementsui.wizardmain.Activator
+Bundle-Activator: org.palladiosimulator.measurementsui.wizard.main.Activator
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.ui.ide,
@@ -26,8 +26,8 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Automatic-Module-Name: de.uni-stuttgart.enpro.wizard
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
-Export-Package: org.palladiosimulator.measurementsui.wizardmain,
- org.palladiosimulator.measurementsui.wizardmain.handlers
+Export-Package: org.palladiosimulator.measurementsui.wizard.handlers,
+ org.palladiosimulator.measurementsui.wizard.main
 Import-Package: emptymeasuringpoints,
  emptymeasuringpointswizard,
  org.palladio.simulator.measurementsui.selectmeasurements.emptyview,
