+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/handlers/contentprovider/AdditionalMeasuringpointContentProvider.java
@@ -1,55 +1,46 @@
-package org.palladiosimulator.measurementsui.wizardmain.handlers;
-
-import java.util.LinkedList;
+package org.palladiosimulator.measurementsui.wizard.handlers.contentprovider;
 
 import org.eclipse.jface.viewers.ITreeContentProvider;
 import org.palladiosimulator.measurementsui.wizardmodel.pages.MeasuringPointSelectionWizardModel;
 
 /**
+ * A content provider for the second step of the measuringpoint creation wizard pages
  * 
  * @author Domas Mikalkinas
  *
  */
-public class MeasuringPointsContentProvider implements ITreeContentProvider {
-
+public class AdditionalMeasuringpointContentProvider implements ITreeContentProvider {
     private MeasuringPointSelectionWizardModel measuringPointWizardModel;
 
     /**
-     * constructor which sets the needed wizard model
+     * constructor with arguments, which sets the needed wizard model
      * 
      * @param measuringPointWizardModel
      *            the needed wizard model
      */
-    public MeasuringPointsContentProvider(MeasuringPointSelectionWizardModel measuringPointWizardModel) {
+    public AdditionalMeasuringpointContentProvider(MeasuringPointSelectionWizardModel measuringPointWizardModel) {
         this.measuringPointWizardModel = measuringPointWizardModel;
     }
 
     @Override
     public Object[] getElements(Object inputElement) {
-        return measuringPointWizardModel.getAllSecondPageObjects();
+
+        return measuringPointWizardModel.getAllAdditionalModels();
     }
 
     @Override
     public Object[] getChildren(Object parentElement) {
-        if (parentElement instanceof LinkedList) {
-            return ((LinkedList<?>) parentElement).toArray();
-        }
         return new Object[0];
     }
 
     @Override
     public Object getParent(Object element) {
-        if (element instanceof LinkedList && !((LinkedList<?>) element).isEmpty()) {
-
-            return element;
-
-        }
         return null;
     }
 
     @Override
     public boolean hasChildren(Object element) {
-        return element instanceof LinkedList;
+        return false;
     }
 
 }
