+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/handlers/contentprovider/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains the ContentProvider classes that are used in the wizard.
+ * @author Ba
+ *
+ */
+package org.palladiosimulator.measurementsui.wizard.handlers.contentprovider;
