+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/handlers/labelprovider/MeasuringPointsLabelProvider.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.measurementsui.wizardmain.handlers;
+package org.palladiosimulator.measurementsui.wizard.handlers.labelprovider;
 
 import java.util.LinkedList;
