+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/handlers/package-info.java
@@ -2,4 +2,4 @@
  * This package contains all the handler classes which are used for the wizard pages.
  * @author Ba
  */
-package org.palladiosimulator.measurementsui.wizardmain.handlers;
+package org.palladiosimulator.measurementsui.wizard.handlers;
