+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/main/Activator.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.measurementsui.wizardmain;
+package org.palladiosimulator.measurementsui.wizard.main;
 
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
