+++ org.palladiosimulator.measurementsui/bundles/org.palladiosimulator.measurementsui.wizard/src/org/palladiosimulator/measurementsui/wizard/pages/package-info.java
@@ -2,4 +2,4 @@
  * This package contains the objects which model the pages of the wizard.
  * @author Ba
  */
-package org.palladiosimulator.measurementsui.wizardpages;
+package org.palladiosimulator.measurementsui.wizard.pages;
